Docint/requirements.txt
docling>=2.0.0
pdf2image
openai
python-dotenv
pillow